actualizando modelos

+++ core/stocks/models.py
@@ -5,7 +5,8 @@ from accounts.models import Client
 class Company(models.Model):
     ruc = models.CharField(max_length=11, primary_key=True)
     acronym = models.CharField(max_length=12)
-    lastest_price = models.FloatField()
+    lastest_price = models.DecimalField(max_digits=7,decimal_places=2)
+
     stocks_for_client = models.ManyToManyField(Client, through="Portfolio")
 
 
@@ -21,7 +22,7 @@ class Order(models.Model):
     client_dni = models.ForeignKey(Client, on_delete=models.CASCADE)
     company_ruc = models.ForeignKey(Company, on_delete=models.CASCADE)
     quantity = models.IntegerField()
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=7,decimal_places=2)
     transaction_type = models.BooleanField()
 
 
@@ -32,4 +33,4 @@ class IncompleteOrder(models.Model):
 
 class CompleteOrder(models.Model):
     order_id = models.ForeignKey(Order,on_delete = models.CASCADE)
-    price_per_stock = models.FloatField()
+    price_per_stock = models.DecimalField(max_digits=7,decimal_places=2)
